+++ pg_documentdb/sql/udfs/projection/bson_projection--0.102-0.sql
@@ -29,6 +29,12 @@ CREATE OR REPLACE FUNCTION __API_SCHEMA_INTERNAL_V2__.bson_dollar_project_find(d
  IMMUTABLE PARALLEL SAFE CALLED ON NULL INPUT
 AS 'MODULE_PATHNAME', $function$bson_dollar_project_find$function$;
 
+CREATE OR REPLACE FUNCTION __API_SCHEMA_INTERNAL_V2__.bson_dollar_project_find(document __CORE_SCHEMA_V2__.bson, pathSpec __CORE_SCHEMA_V2__.bson, querySpec __CORE_SCHEMA_V2__.bson, letVariableSpec __CORE_SCHEMA_V2__.bson, collationString text)
+ RETURNS __CORE_SCHEMA_V2__.bson
+ LANGUAGE c
+ IMMUTABLE PARALLEL SAFE CALLED ON NULL INPUT
+AS 'MODULE_PATHNAME', $function$bson_dollar_project_find$function$;
+
 CREATE OR REPLACE FUNCTION __API_CATALOG_SCHEMA__.bson_dollar_add_fields(document __CORE_SCHEMA__.bson, pathSpec __CORE_SCHEMA__.bson)
  RETURNS __CORE_SCHEMA__.bson
  LANGUAGE c